+++ printf/src/ft_count_digits.c
@@ -6,13 +6,13 @@
 /*   By: ppatil <ppatil@student.42.us.org>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/09/28 18:02:58 by ppatil            #+#    #+#             */
-/*   Updated: 2016/09/28 18:08:03 by ppatil           ###   ########.fr       */
+/*   Updated: 2017/04/18 16:49:12 by ppatil           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_math.h"
 
-size_t	ft_count_digits(long n, size_t base)
+size_t	ft_count_digits(long long n, size_t base)
 {
 	size_t	digit_count;
 
@@ -22,7 +22,7 @@ size_t	ft_count_digits(long n, size_t base)
 	while (n)
 	{
 		++digit_count;
-		n /= (long)base;
+		n /= (long long)base;
 	}
 	if (!digit_count)
 		return (1);
